move cp-related code from sfa_v1.py into sfa_cp.py
